Remove "file://" component references from public docs
A benefit of smithy is that you don't need the source code of every
single connector in order to run it. If our docs have workflows that only reference file:// type components, then our users can't benefit from this.
This gets increasingly frustrating as the repo gets larger, especially since we're vendoring components.
This commit rewrites docs to replace file:// type references to
components with ghcr:// type references to components.
It also slightly cleans up the 'how to write a workflow' piece of the
open source docs explaining the difference between file:// component
references used mostly for development and 'ghcr://' type component
references used for day to day operations. It also documents how the
component registry works and how to use it.